refactor: rename ui/registry.ts → ui/loader.ts to match pi/loader.ts
